Update UI and Improve Setup
